Rename listEditable to capabilitiesEditable
