+++ src/components/CapabilitiesEditable/index.tsx
@@ -1,0 +1,1 @@
+export { CapabilitiesEditableContainer as CapabilitiesEditable } from "./container";
