--- src/components/ListEditable/index.tsx
@@ -1,1 +1,0 @@
-export { ListEditableContainer as ListEditable } from "./container";
