--- src/components/ListEditable/interface.tsx
@@ -1,6 +1,0 @@
-export interface IListEditableProps {
-  setList: (newValue: string | number) => void;
-  list?: string[];
-  title: string;
-  onFinish?: () => void;
-}
